+++ db/migrate/20190423135949_add_featured_to_broadcasts.rb
@@ -1,0 +1,6 @@
+class AddFeaturedToBroadcasts < ActiveRecord::Migration[6.0]
+  def change
+    add_column :broadcasts, :featured, :boolean, null: false, default: false
+    add_index :broadcasts, :featured
+  end
+end